back/requirements.txt
python-jose[cryptography]~=3.3.0
bcrypt~=4.1.2
fastapi~=0.109.0
# passlib~=1.7.4
pydantic~=1.10.14
uvicorn
python-multipart
sqlmodel~=0.0.14